LayoutTests/imported/w3c/web-platform-tests/screen-orientation/onchange-event-subframe-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'window.screen.orientation.addEventListener')

Harness Error (FAIL), message = Test named 'Test subframes receive orientation change events' specified 1 'cleanup' function, and 1 failed.

FAIL Test subframes receive orientation change events promise_test: Unhandled rejection with value: object "TypeError: document.documentElement.requestFullscreen is not a function. (In 'document.documentElement.requestFullscreen()', 'document.documentElement.requestFullscreen' is undefined)"
